distributed-task-manager/services/users-service/requirements.txt
flask
flask-pymongo
PyJWT
werkzeug
gunicorn
